requirements.txt
flask>=3.0
azure-monitor-opentelemetry>=1.6
azure-identity>=1.15
azure-cosmos>=4.7
azure-storage-blob>=12.19
openai>=1.40
python-dotenv>=1.0
gunicorn>=22.0
